fix: named exports for v5 compatibility
The change from module.exports to export default no longer allows you to use named imports. This would be a breaking change, but it's also incompatible with the current type declarations.

This could be fixed using commonjs module and 'export =' in TypeScript, but this is incompatible with Rollup. The simplest answer seems to be just explicitly using named exports based on the type declarations as well as including the default export.